interconnect: qcom: icc-rpm: Add AB/IB calculations coefficients
Presumably due to the hardware being so complex, some nodes (or busses)
have different (usually higher) requirements for bandwidth than what
the usual calculations would suggest.

Looking at the available downstream files, it seems like AB values are
adjusted per-bus and IB values are adjusted per-node.
With that in mind, introduce percentage-based coefficient struct members
and use them in the calculations.

One thing to note is that downstream does (X%)*AB and IB/(Y%) which
feels a bit backwards, especially given that the divisors for IB turn
out to always be 25, 50, 200 making this a convenient conversion to 4x,
2x, 0.5x.. This commit uses the more sane, non-inverse approach.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>